requirements.txt
asgiref==3.5.2
Django==3.2.7
django-cors-headers==3.10.0
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
phonenumbers==8.12.50
psycopg2-binary==2.9.2
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
